Tarefa5_2022li1g016

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo :: Jogo -> Int -> Jogo

A função deslizaJogo adiciona uma linha, pseudo-aleatória, no limite superior do mapa ao mesmo tempo que elemina a linha mais abaixo do mapa, assim mantendo o tamanho do mapa.

deslizaJogo :: Jogo -> Int -> Jogo
deslizaJogo (Jogo (Jogador(x,y)) (Mapa l ((terreno, obstaculos):t))) n = (Jogo (Jogador(x,y+1)) (estendeMapa (Mapa l (init ((terreno, obstaculos):t))) n))

Exemplos de utilização:

>>> deslizaJogo (Jogo (Jogador (1,1)) (Mapa 3 [(Estrada (-2), [Carro, Carro, Nenhum]), (Rio (-1), [Nenhum, Tronco, Tronco])])) 1 = Jogo (Jogador (1,2)) (Mapa 3 [(Estrada (-2),[Nenhum,Carro,Carro]),(Estrada (-2),[Carro,Carro,Nenhum])])